Merged PR 80873: Change CKAN to deployment and add Kubernetes Cronjob
Archiving of messages from the EventGrid queues, were set to be saved on a disk mounted to the CKAN pods. This has since been uncovered to be a bad and inefficient practice. Thus we are moving the native cronjob from within the container, to be a seperate Kubernetes resource. This gives more transparity in what is running, and an easier way to find logs for the cronjob runs.

Messages will now be backed up on Azure Blob storage, instead of on mounted disks. This gives us the backup feature as well as keeping the containers clean.

Thus removing the needed mounted disks, means that CKAN can now be a simple Deployment instead of a StatefulSet.

A new resource Cronjobs has been added as part of the base manifests, and will be deployed alongside the CKAN webapp.

Changes:
 - Cleaned up dependencies for using statefulset and changed too deployment.
 - Changed deploy-steps for Azure Pipelines to remove and deploy the deployment and cronjob.
 - Updated environment variables and secrets to include the SPN secrets and Tenant Ids - to reflect usage of AD authentication instead of SAS tokens that needs to be updated regularly.
 - Removed volumeclaim for CKAN as no volume is needed. A PR will be submitted for the k8s team to remove the attached disk and StorageClass in their own repos.
 - Removed shell script and cronjob from ckan configs manifests.